Extract template dependencies
